requirements.txt
astropy==4.3.1
bandit==1.7.0
cmdstanpy==0.9.5
convertdate==2.3.2
cycler==0.10.0
Cython==0.29.24
ephem==4.0.0.2
fbprophet==0.7.1
gitdb==4.0.7
GitPython==3.1.18
hijri-converter==2.2.0
holidays==0.11.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.4.3
numpy==1.21.2
P4J==0.27
pandas==1.3.2
patsy==0.5.1
pbr==5.6.0
Pillow==8.3.1
prophet==1.0.1
psycopg2-binary==2.9.1
pyerfa==2.0.0
PyMeeus==0.5.11
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scipy==1.7.1
setuptools-git==1.2
six==1.16.0
smmap==4.0.0
statsmodels==0.12.2
stevedore==3.4.0
tqdm==4.62.2
ujson==4.1.0
